bdd/requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Admin==1.5.8
requests==2.27.1
pyTelegramBotAPI==4.3.0